Fixes error where PailSpec args are ignored.
Explicitly adds the file format name when creating the PailSpec.

When a Pail is created, the PailFormatFactory checks to see if there is
a PailSpec arg. If the arg is null or if the name of the PailSpec is
null then a new default PailSpec is created and used.

This necesitates a PailSpec name being explicitly set when options are
present, else a default PailSpec is used and the options are lost.