Adds file format parameter to PailSpec creation.
In a PailSpec, the options are tied specifically to the PailFormat used. The PailFormat is the parameter called "name" in the PailSpec constructors.

If using the options parameter, we need to know the PailFormat
used. Not passing in a PailFormat and letting the constructor set a
default format is dangerous, relying on implementation details.
We should always force a PailFormat name to be passed in when a PailSpec is created.